vertez/requirements.txt
numpy==1.23.5
pandas==2.0.1
tensorflow==2.13.0
tensorflow-probability==0.21.0
# arviz==0.14.0
google-cloud-storage>=2.0.0
fsspec>=2023.6.0
gcsfs
ipython